Enable extra warnings, adjust code.
We would like to be able to build with -pedantic -Wcast-qual, the former
because it flags accidental use of non-standard language constructs, the
latter because it is also enabled by LLVM. These are now enabled, and
code is updated to address the warnings raised.

The C standard is bumped from C99 to C11 to account for the fact that we
were already using C11 language features, which now got flagged by
-pedantic. The move to C11 also enables the
checkDeprecatedOrUnsafeBufferHandling warning, which tells us not to use
scanf and to use scanf_s instead, despite scanf_s being optional and not
supported on platforms we support, and despite the use of scanf being
safe. Therefore, this warning is disabled.

+++ modules/mux/include/mux/mux.h
@@ -218,7 +218,7 @@ enum mux_object_id_e {
   mux_object_id_fence = (13 << 16),
   /// @brief Mask for ID values which are always help in the upper 16 bits of
   /// the 32 bit `::mux_id_t`.
-  mux_object_id_mask = 0xffff0000
+  mux_object_id_mask = (int)0xffff0000
 };
 
 /// @brief All possible device types.
@@ -241,7 +241,7 @@ enum mux_device_type_e {
   /// @brief Device is a mysterious custom type.
   mux_device_type_custom = (0x1 << 5),
   /// @brief Mask to match all device types.
-  mux_device_type_all = 0xFFFFFFFF
+  mux_device_type_all = (int)0xFFFFFFFF
 };
 
 /// @brief All possible queue types that can be supported.
